Refactoring to move quantity estimation for the vetter into lib/stack.h so I can use it for stack as well.
